+++ NEWS.md
@@ -4,6 +4,7 @@
 * Added parallel method in `solve_mcsim()`
 * Fixed bug in `heat_check()`
 * Added message of display info in `heat_check()` function
+* Improved `heat_check()` to display the indices based on the setting cut-off
 
 # pksensi 1.1.6