+++ NEWS.md
@@ -1,3 +1,7 @@
+# pksensi 1.1.8
+
+* update the route to load example model
+
 # pksensi 1.1.7
 
 * Suggested to use Rtools40